feat(merge): set a merge up beside the results, not above them
Merge configuration sat in a band between the filters and the chart, so
building one pushed the results down the page and reading the results meant
scrolling away from the thing that shaped them. It also read as a section of
the page rather than something you do to the query.

It is a column beside the results now, sticky while you scroll, and the way
in sits next to the control that runs it — adding a second query is a thing
you do to the query you are about to run.

Renamed from strip to panel, since that is what it is.
