+++ packages/frontend/src/components/Explorer/ExplorerHeader/index.tsx
@@ -10,6 +10,7 @@ import {
     selectUnsavedChartVersion,
     useExplorerSelector,
 } from '../../../features/explorer/store';
+import { MergeToggleButton } from '../../../features/mergeQuery/components/MergeToggleButton';
 import useDashboardStorage from '../../../hooks/dashboard/useDashboardStorage';
 import { useExplorerQuery } from '../../../hooks/useExplorerQuery';
 import { getExplorerUrlFromCreateSavedChartVersion } from '../../../hooks/useExplorerRoute';
@@ -174,6 +175,8 @@ const ExplorerHeader: FC = memo(() => {
                     queryWarnings.length > 0 && (
                         <QueryWarnings queryWarnings={queryWarnings} />
                     )}
+
+                <MergeToggleButton />
 
                 <RefreshButton size="xs" />
