+++ packages/frontend/src/features/mergeQuery/components/MergeSetupPanel.module.css
@@ -1,0 +1,22 @@
+/*
+ * An inline column beside the results rather than a band above them: the merge
+ * is the shape of the query, so it stays legible while you read what it
+ * produced instead of pushing the chart down the page.
+ */
+.panel {
+    width: 340px;
+    flex: none;
+    align-self: flex-start;
+    position: sticky;
+    top: var(--mantine-spacing-md);
+    max-height: calc(100vh - var(--page-content-height-offset, 8rem));
+    overflow-y: auto;
+}
+
+@media (max-width: 1200px) {
+    .panel {
+        width: 100%;
+        position: static;
+        max-height: none;
+    }
+}
